Refactor new_dat to dat
Going to pull to master, so this should all be the default, not "new"
