+++ src/dat_analysis/__init__.py
@@ -1,5 +1,5 @@
 # from .dat_object.make_dat import get_dat, get_dats
-from .new_dat.new_dat_util import get_local_config
-from .new_dat.dat_hdf import get_dat, get_dat_from_exp_filepath
+from .dat.new_dat_util import get_local_config
+from .dat.dat_hdf import get_dat, get_dat_from_exp_filepath
 
 config = get_local_config()
