+++ tests/new_dat/test_dat_hdf.py
@@ -15,8 +15,8 @@
 #     test_filename = 'dat1.h5'
 #     with h5py.File(os.path.join(config['loading']['path_to_save_directory'], test_filename), 'w') as f:
 from unittest import TestCase
-from dat_analysis.new_dat.dat_hdf import DatHDF, get_dat, get_dat_from_exp_filepath
-import dat_analysis.new_dat.new_dat_util as ndu
+from dat_analysis.dat.dat_hdf import DatHDF, get_dat, get_dat_from_exp_filepath
+import dat_analysis.dat.new_dat_util as ndu
 import os
 from .helper import setup_test_config
